move mock_calibration_dir in utils
